% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204136R.R
\name{FSR}
\alias{FSR}
\title{Fisher scoring algorithm for cauchy(theta) using R.}
\usage{
FSR(x, theta, method)
}
\arguments{
\item{x}{the data of cauchy distribution}

\item{theta}{the initial value}

\item{method}{method=1 for one step iteration and method=2 for multi-step iteration}
}
\value{
the estimation of param theta
}
\description{
Using three input to get the result of Fisher scoring algorithm of cauchy distribution.
}
\examples{
\dontrun{
x<-rcauchy(200,location=1)
theta<-median(x)
thetahat<-FSR(x,theta,1)
print(thetahat)
}
}
