\name{SA24204136-package}
\alias{SA24204136-package}
\docType{package}
\title{This package provides several functions for Newton-Raphson algorithm and Fisher scoring algorithm}
\description{R and Rcpp versions of Newton-Raphson algorithm and Fisher scoring algorithm}
\author{
Sheng Liu
Maintainer:Sheng Liu <824586948@qq.com>
}
\keyword{package}
